server/requirements.txt
Django>=5.2.7
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.1
django-cors-headers>=4.9.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
